Make loader address arithmetic 64-bit-safe (pointer hygiene, stage 3c)
Two loader sites cast pointers to 32 bits and would misbehave on x64:

- ImageLoader::bLoadHeapAllocation tested whether an arbitrary pointer
  lies inside the load heap by truncating pv and the heap base/end to
  uint32 and comparing. If the heap is mapped above 4GB, the truncated
  absolute-address comparison gives the wrong answer (a real correctness
  bug, not just a warning). Widened the comparison to uintptr.

- CTextureManager::CreatePackedImageDirectory computes a texture's
  u4VMOffset as pSurface - VM-base by casting both pointers to uint32
  first. The offset itself fits in uint32 (the VM block is < 4GB), so
  widen only the two pointers to uintptr and keep the uint32 result via
  an explicit narrow. Same for the bump-map path.

Left TexturePackSurface.cpp untouched: its (uint32)/(int) casts are all
on uint8/size values (u1XOrg, u1Count, u4ParentSize), not pointers, and
TextureManager's abs((int32)ept_surface - (int32)ept) is an enum
(ETexturePackTypes) difference, not pointer arithmetic.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Third slice of the uint32->uintptr
sweep after VirtualMem (ba41823) and FastHeap (3825cd3).

+++ jp2_pc/Source/Lib/Loader/ImageLoader.hpp
@@ -209,8 +209,8 @@ public:
 	)
 	//*************************************
 	{
-		return ( ((uint32)pv < ((uint32)fhImageLoad.uNumBytesUsed() + (uint32)fhImageLoad.pvGetBase())) &&
-		((uint32)pv >= (uint32)fhImageLoad.pvGetBase()) );
+		return ( ((uintptr)pv < (fhImageLoad.uNumBytesUsed() + (uintptr)fhImageLoad.pvGetBase())) &&
+		((uintptr)pv >= (uintptr)fhImageLoad.pvGetBase()) );
 	}
